requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.45.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
ruff==0.4.5
black==24.4.2
mypy==1.10.0
pytest==8.2.1
pytest-asyncio==0.23.6
coverage==7.5.3
